change: Move the new nested classes to their own files.
