+++ Yafc.Model/Model/ProductionTable.cs
@@ -17,7 +17,7 @@ public struct ProductionTableFlow(IObjectWithQuality<Goods> goods, float amount,
     public IProductionLink? link = link;
 }
 
-public class ProductionTable : ProjectPageContents, IComparer<ProductionTableFlow>, IElementGroup<RecipeRow> {
+public sealed partial class ProductionTable : ProjectPageContents, IComparer<ProductionTableFlow>, IElementGroup<RecipeRow> {
     private static readonly ILogger logger = Logging.GetLogger<ProductionTable>();
     [SkipSerialization] public Dictionary<IObjectWithQuality<Goods>, IProductionLink> linkMap { get; } = [];
     List<RecipeRow> IElementGroup<RecipeRow>.elements => recipes;
@@ -771,148 +771,6 @@ match:
         float amt2 = y.goods.target.fluid != null ? y.amount / 50f : y.amount;
 
         return amt1.CompareTo(amt2);
-    }
-
-    /// <summary>
-    /// An implicitly-created link, used to ensure synthetic recipes are properly connected to the rest of the production table.
-    /// </summary>
-    /// <param name="goods">The linked <see cref="ObjectWithQuality{T}"/>.</param>
-    /// <param name="owner">The <see cref="ProductionTable"/> that owns this link.</param>
-    /// <param name="displayLink">The ordinary link that caused the creation of this implict link, and the link that will be displayed if the
-    /// user requests the summary for this link.</param>
-    private class ImplicitLink(ObjectWithQuality<Goods> goods, ProductionTable owner, ProductionLink displayLink) : IProductionLink {
-        /// <summary>
-        /// Always <see cref="LinkAlgorithm.Match"/>; implicit links never allow over/under production.
-        /// </summary>
-        public LinkAlgorithm algorithm => LinkAlgorithm.Match;
-
-        public ObjectWithQuality<Goods> goods { get; } = goods;
-
-        /// <summary>
-        /// Always 0; implicit links never request additional production or consumption.
-        /// </summary>
-        public float amount => 0;
-
-        public HashSet<IRecipeRow> capturedRecipes { get; } = [];
-        public int solverIndex { get; set; }
-        public float linkFlow { get; set; }
-        public float notMatchedFlow { get; set; }
-
-        public ProductionTable owner { get; } = owner;
-        public ProductionLink.Flags flags { get; set; }
-
-        /// <summary>
-        /// Always empty; implicit links never have warnings.
-        /// </summary>
-        public IEnumerable<string> LinkWarnings { get; } = [];
-
-        public ProductionLink DisplayLink { get; } = displayLink;
-    }
-
-    /// <summary>
-    /// A recipe for allowing the solver to treat one quality science pack as (Quality.level+1) normal science packs
-    /// </summary>
-    /// <param name="pack">The normal-quality pack produced by this recipe.</param>
-    /// <param name="quality">The quality of the recipe's input pack.</param>
-    /// <param name="productLink">The (genuine) link for the output pack.</param>
-    /// <param name="ingredientLink">The (implicit) link for the input pack.</param>
-    private class ScienceDecomposition(Goods pack, Quality quality, IProductionLink productLink, ImplicitLink ingredientLink) : IRecipeRow {
-        public ObjectWithQuality<EntityCrafter>? entity => null;
-
-        public ObjectWithQuality<Goods>? fuel => null;
-
-        /// <summary>
-        /// Always 0; the solver must scale this recipe to match the available quality packs.
-        /// </summary>
-        public float fixedBuildings => 0;
-
-        /// <inheritdoc/>
-        public IEnumerable<SolverIngredient> IngredientsForSolver => [new SolverIngredient(pack.With(quality), 1, ingredientLink, 0)];
-        /// <inheritdoc/>
-        public IEnumerable<SolverProduct> ProductsForSolver => [new SolverProduct(pack.With(Quality.Normal), quality.level + 1, productLink, 0, null)];
-
-        public double recipesPerSecond { get; set; }
-        public RecipeParameters parameters { get; set; } = RecipeParameters.Empty;
-        public RecipeLinks links => new() { ingredients = [null] };
-
-        /// <summary>
-        /// Always 1; the solver needs something non-zero to calculate <see cref="fixedBuildings"/>.
-        /// </summary>
-        public float RecipeTime => 1;
-
-        /// <inheritdoc/>
-        public string SolverName => $"Convert {pack.name} from {quality.name}";
-
-        public double BaseCost => 0;
-
-        public void GetModulesInfo((float recipeTime, float fuelUsagePerSecondPerBuilding) recipeParams, EntityCrafter entity, ref ModuleEffects effects, ref UsedModule used) => throw new NotImplementedException();
-
-        public bool FindLink(IObjectWithQuality<Goods> goods, [MaybeNullWhen(false)] out IProductionLink link) {
-            if (ingredientLink.goods == goods) {
-                link = ingredientLink;
-                return true;
-            }
-            if (productLink.goods == goods) {
-                link = productLink;
-                return true;
-            }
-            link = null;
-            return false;
-        }
-
-        /// <summary>
-        /// Always null; this does not represent a user-visible <see cref="RecipeRow"/>.
-        /// </summary>
-        RecipeRow? IRecipeRow.RecipeRow => null;
-    }
-
-    /// <summary>
-    /// A recipe that appears in the production table, with additional context allowing it to find and obey <see cref="ImplicitLink"/>s.
-    /// </summary>
-    /// <param name="row">The underlying <see cref="RecipeRow"/>.</param>
-    /// <param name="extraLinks">The <see cref="Dictionary{TKey, TValue}"/> that will eventually the contain extra links this recipe needs to
-    /// consider. This may be incomplete at the time of construction, provided it is complete before the first call to <see cref="FindLink"/>.
-    /// </param>
-    private class GenuineRecipe(RecipeRow row, Dictionary<(ProductionTable, IObjectWithQuality<Goods>), IProductionLink> extraLinks) : IRecipeRow {
-        /// <summary>
-        /// Check both <see cref="row"/> and <see cref="extraLinks"/> for a link corresponding to <paramref name="goods"/>.
-        /// </summary>
-        public bool FindLink(IObjectWithQuality<Goods> goods, [MaybeNullWhen(false)] out IProductionLink link) {
-            // Get the genuine user-visible link, if present.
-            _ = row.FindLink(goods, out var genuineLink);
-
-            // Search for a more nested synthetic link
-            ProductionTable? linkOwner = row.linkRoot;
-            while (linkOwner != genuineLink?.owner) {
-                if (extraLinks.TryGetValue((linkOwner!, goods), out link)) { // null-forgiving: genuineLink?.owner is null, linkOwner, or an ancestor of linkOwner.
-                    // There's a more-nested synthetic link, so return it instead of the one the project knows about.
-                    return true;
-                }
-                linkOwner = linkOwner!.owner.ownerObject as ProductionTable;
-            }
-
-            // Return the genuine link
-            link = genuineLink;
-            return link != null;
-        }
-
-        // Pass all remaining calls through to the underlying RecipeRow.
-        public ObjectWithQuality<EntityCrafter>? entity => row.entity;
-        public ObjectWithQuality<Goods>? fuel => row.fuel;
-        public float fixedBuildings => row.fixedBuildings;
-        public double recipesPerSecond { get => row.recipesPerSecond; set => row.recipesPerSecond = value; }
-        public float RecipeTime => ((IRecipeRow)row).RecipeTime;
-        public RecipeParameters parameters { get => row.parameters; set => row.parameters = value; }
-        public RecipeLinks links => row.links;
-        public IEnumerable<SolverIngredient> IngredientsForSolver => ((IRecipeRow)row).IngredientsForSolver;
-        public IEnumerable<SolverProduct> ProductsForSolver => ((IRecipeRow)row).ProductsForSolver;
-        public string SolverName => ((IRecipeRow)row).SolverName;
-        public double BaseCost => ((IRecipeRow)row).BaseCost;
-
-        public void GetModulesInfo((float recipeTime, float fuelUsagePerSecondPerBuilding) recipeParams, EntityCrafter entity, ref ModuleEffects effects, ref UsedModule used)
-            => row.GetModulesInfo(recipeParams, entity, ref effects, ref used);
-
-        RecipeRow? IRecipeRow.RecipeRow => row;
     }
 }
